ecommerce/requirements.txt
accessible-pygments==0.0.5
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.16
amplitude-analytics==1.1.4
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.4.0
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.2.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.1
aws-sam-translator==1.91.0
aws-xray-sdk==2.14.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.10.0
blis==0.7.11
boto3==1.34.51
botocore==1.34.162
cachetools==5.5.0
catalogue==2.0.10
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
cfn-lint==1.10.3
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.18.1
colorama==0.4.6
colorlog==6.7.0
confection==0.1.5
coverage==7.6.2
crispy-bootstrap5==0.7
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect2==0.7.0
cymem==2.0.8
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
Django==5.1.2
django-allauth==0.51.0
django-allow-cidr==0.7.1
django-celery-beat==2.5.0
django-cleanup==6.0.0
django-colorfield==0.9.0
django-cors-headers==3.10.0
django-coverage-plugin==2.0.4
django-crispy-forms==1.14.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==21.1
django-guardian==2.4.0
django-health-check==3.17.0
django-invitations==2.0.0
django-js-asset==2.2.0
django-linear-migrations==2.10.0
django-model-utils==4.5.1
django-mptt==0.16.0
django-multitenant==3.2.1
django-nested-admin==4.0.2
django-phonenumber-field==6.0.0
django-polymorphic==3.1.0
django-ratelimit==4.1.0
django-redis==5.3.0
django-rest-passwordreset==1.2.1
django-rest-polymorphic==0.1.10
django-restql==0.15.1
django-ses==4.1.1
django-simple-history==3.7.0
django-storages==1.14
django-stubs==4.2.4
django-stubs-ext==5.0.4
django-test-migrations==1.3.0
django-timezone-field==7.0
django_debug_toolbar==3.8.1
djangorestframework==3.15.2
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.14.3
dnspython==2.6.1
docker==7.1.0
docutils==0.19
drf-nested-routers==0.93.4
drf-rw-serializers==1.4.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2021.12.13
et-xmlfile==1.1.0
execnet==2.1.1
executing==2.0.1
factory_boy==3.3.1
Faker==30.3.0
favicon==0.7.0
filelock==3.15.4
flake8==7.1.1
fonttools==4.53.1
frozenlist==1.4.1
gevent==23.9.1
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.63.0
gql==3.5.0
graphql-core==3.2.3
graphviz==0.20.3
greenlet==3.0.0
h11==0.14.0
hiredis==3.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
identify==2.6.0
idna==3.7
imagesize==1.4.1
IMAPClient==2.3.1
inflection==0.5.1
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jira==3.8.0
jmespath==1.0.1
joserfc==1.0.0
jsondiff==2.2.0
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2024.10.1
# Editable install with no version control (k_cli==1.0)
-e /Users/marcusmartinez/Desktop/k_cli
kombu==5.3.7
langchain==0.2.3
langchain-aws==0.1.7
langchain-community==0.2.4
langchain-core==0.2.33
langchain-experimental==0.0.60
langchain-text-splitters==0.2.2
langcodes==3.4.0
langdetect==1.0.9
langfuse==2.37.0
langsmith==0.1.100
language_data==1.2.0
lazy-object-proxy==1.10.0
libcst==1.4.0
livereload==2.7.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergepythonclient==1.0.13
moto==5.0.10
mpmath==1.3.0
msal==1.28.0
multidict==6.0.5
multipart==0.2.5
murmurhash==1.0.10
mypy==1.5.1
mypy-extensions==1.0.0
mypy_baseline==0.6.1
neo4j==5.23.1
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.2
orjson==3.10.7
oscrypto==1.3.0
packaging==24.1
parso==0.8.4
pathable==0.4.3
pathspec==0.12.1
pexpect==4.9.0
phonenumberslite==8.13.43
pillow==10.3.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
pre-commit==3.7.1
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure_eval==0.2.3
py-partiql-parser==0.5.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.12.1
pycountry==22.3.5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydata-sphinx-theme==0.15.4
pydot==1.4.2
pyflakes==3.2.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pylint==3.2.6
pyparsing==3.1.1
pypdf==4.3.1
pyPEG2==2.15.2
pypng==0.20220715.0
pytest==8.3.3
pytest-cov==3.0.0
pytest-django==4.9.0
pytest-factoryboy==2.7.0
pytest-mock==3.14.0
pytest-recording==0.13.2
pytest-split==0.9.0
pytest-sugar==0.9.7
pytest-testmon==1.4.5
pytest-xdist==3.3.1
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.2
qrcode==7.4.2
redis==4.5.4
referencing==0.35.1
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.3
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
ruff==0.0.292
s3transfer==0.10.2
sentry-sdk==1.40.3
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinx-book-theme==1.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-confluencebuilder==1.9.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
strawberry-graphql==0.242.0
strawberry-graphql-django==0.48.0
StrEnum==0.4.15
svglib==1.5.1
sympy==1.13.2
tenacity==8.5.0
termcolor==2.4.0
text-unidecode==1.3
thinc==8.2.5
tinycss2==1.3.0
tldextract==5.1.1
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.4
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240808
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
vcrpy==6.0.1
vine==5.1.0
virtualenv==20.26.3
wasabi==1.1.3
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
Werkzeug==2.3.8
whitenoise==5.3.0
wrapt==1.16.0
xhtml2pdf==0.2.16
xmltodict==0.13.0
yarl==1.9.4
zipp==3.20.0
zope.event==4.6
zope.interface==5.4.0
